#821 [REFACTOR] shared/component/layout 분류 완료
