--- src/components/AppBar/index.js
@@ -1,3 +1,0 @@
-export { default as AppBar } from './AppBar/AppBar';
-export { default as BackAppBar } from './BackAppBar/BackAppBar';
-export { default as CloseAppBar } from './CloseAppBar/CloseAppBar';
