--- src/components/Navbar/index.js
@@ -1,1 +1,0 @@
-export { default as Navbar } from './Navbar';
