--- src/components/Sidebar/index.js
@@ -1,3 +1,0 @@
-export { default as Sidebar } from './Sidebar';
-
-export { default as MenuList } from './MenuList/MenuList';
