+++ src/components/index.js
@@ -1,4 +1,3 @@
-export * from './AppBar';
 export * from './BoardBar';
 export * from './BoardCard';
 export * from './Card';
@@ -9,8 +8,6 @@ export * from './DropDownMenu';
 export * from './Fieldset';
 export * from './Filter';
 export * from './Flex';
-export * from './Footer';
-export * from './Header';
 export * from './Icon';
 export * from './Input';
 export * from './InputBar';
@@ -21,7 +18,6 @@ export * from './MainPageListItem';
 export * from './Margin';
 export * from './MenuIcon';
 export * from './Modal';
-export * from './Navbar';
 export * from './Portal';
 export * from './NoticeBar';
 export * from './PostBar';
@@ -32,6 +28,5 @@ export * from './ReviewDownload';
 export * from './Search';
 export * from './ServerErrorFallback';
 export * from './ShimmerEffect';
-export * from './Sidebar';
 export * from './StageDots';
 export * from './Toast';
