+++ src/pages/AboutPage/AboutPage.jsx
@@ -1,11 +1,11 @@
 import { Button } from '@/pages/SnoroseVerifyPage';
 
+import { BackAppBar } from '@/shared/component';
 import {
   Accordion,
   AccordionListItem,
   AccordionTag,
 } from '@/components/Accordion';
-import { BackAppBar } from '@/components/AppBar';
 import { Icon } from '@/components/Icon';
 
 import {
