+++ src/pages/BoardPage/BoardPage.jsx
@@ -1,6 +1,7 @@
 import { useNavigate } from 'react-router-dom';
 
-import { BoardBar, Header, Icon, Search } from '@/components';
+import { Header } from '@/shared/component';
+import { BoardBar, Icon, Search } from '@/components';
 import { BOARD_MENUS } from '@/constants';
 
 import styles from './BoardPage.module.css';
