+++ src/pages/ExamReviewEditPage/ExamReviewEditPage.jsx
@@ -6,8 +6,8 @@ import { editReviewDetail } from '@/apis';
 
 import { useToast } from '@/hooks';
 
-import { ActionButton, Textarea } from '@/shared/component';
-import { CloseAppBar } from '@/components/AppBar';
+import { ActionButton, CloseAppBar, Textarea } from '@/shared/component';
+
 import {
   CategoryButton,
   CategoryFieldset,
