+++ src/pages/MyPage/pages/ActivityPage/ActivityPage.jsx
@@ -4,7 +4,8 @@ import { ErrorBoundary } from 'react-error-boundary';
 import { QueryErrorResetBoundary } from '@tanstack/react-query';
 
 import { useScrollRestoration } from '@/hooks';
-import { BackAppBar, FetchLoading } from '@/components';
+import { BackAppBar } from '@/shared/component';
+import { FetchLoading } from '@/components';
 import { Posts, PostsErrorFallback } from '@/pages/MyPage/pages/ActivityPage';
 import { ACTIVITIES } from '@/constants';
