--- src/shared/component/button/Submit/index.js
@@ -1,1 +1,0 @@
-export { default as Submit } from './Submit';
