--- src/shared/component/button/WriteButton/index.js
@@ -1,1 +1,0 @@
-export { default as WriteButton } from './WriteButton';
