+++ src/shared/component/index.js
@@ -9,3 +9,13 @@ export { default as Input } from './input/Input/Input';
 export { default as InputPassword } from './input/InputPassword/InputPassword';
 export { default as Textarea } from './input/Textarea/Textarea';
 export { default as SpecialInput } from './input/Input/SpecialInput';
+
+// layout
+export { default as AppBar } from './layout/AppBar/AppBar';
+export { default as BackAppBar } from './layout/BackAppBar/BackAppBar';
+export { default as CloseAppBar } from './layout/CloseAppBar/CloseAppBar';
+export { default as Footer } from './layout/Footer/Footer';
+export { default as Header } from './layout/Header/Header';
+export { default as MenuList } from './layout/Sidebar/MenuList/MenuList';
+export { default as Navbar } from './layout/Navbar/Navbar';
+export { default as Sidebar } from './layout/Sidebar/Sidebar';
