--- src/shared/component/input/InputPassword/index.js
@@ -1,1 +1,0 @@
-export { default as InputPassword } from './InputPassword';
