+++ src/shared/component/layout/CloseAppBar/CloseAppBar.module.css
